Replace NoOpTelemetry with CKTelemetryHandler.Sink
Removed the NoOpTelemetry class and replaced its usage in TelemetryFactory with CKTelemetryHandler.Sink. Added the Sink implementation to CKTelemetryHandler to serve as a no-op handler for metrics and recorder interfaces.
